# docs/swagger.yaml
basePath: /
definitions:
  models.Car:
    properties:
      brand:
        type: string
      id:
        type: integer
      model:
        type: string
      price:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: soberkoder@swagger.io
    name: API Support
  description: This is a sample service for managing cars
  license:
    name: Apache 2.0
  title: Car API
  version: "1.0"
paths:
  /cars:
    post:
      consumes:
      - application/json
      description: Post detials of car corresponding to the input Id
      parameters:
      - description: create car
        in: body
        name: models.Car
        required: true
        schema:
          $ref: '#/definitions/models.Car'
      produces:
      - application/json
      responses: {}
      summary: Post details for a given Id
      tags:
      - cars
  /cars/{Id}:
    delete:
      consumes:
      - application/json
      description: Delete the order corresponding to the input Id
      parameters:
      - description: ID of the car to be deleted
        in: path
        name: Id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete car identified by the given Id
      tags:
      - cars
    get:
      consumes:
      - application/json
      description: Get details of car corresponding to the input Id
      parameters:
      - description: ID of the car
        in: path
        name: Id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Car'
      tags:
      - cars
  /orders:
    get:
      consumes:
      - application/json
      description: Get details of all car
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Car'
      summary: Get details
      tags:
      - cars
swagger: "2.0"